+++ src/app/app.component.ts
@@ -3,7 +3,8 @@ import { Component } from '@angular/core';
 import { Platform } from '@ionic/angular';
 import { SplashScreen } from '@ionic-native/splash-screen/ngx';
 import { StatusBar } from '@ionic-native/status-bar/ngx';
-import { OneSignal } from '@ionic-native/onesignal/ngx';
+import { Notification } from 'rxjs';
+import { NotificationService } from './device-information/notification.service';
 
 @Component({
   selector: 'app-root',
@@ -12,39 +13,10 @@ import { OneSignal } from '@ionic-native/onesignal/ngx';
 export class AppComponent {
   // Add the following to your existing ready fuction.
 
-constructor(private platform: Platform, private oneSignal: OneSignal, private statusBar: StatusBar, private splashScreen: SplashScreen) {
+constructor(private platform: Platform, private statusBar: StatusBar, private splashScreen: SplashScreen, notificationsService: NotificationService) {
   platform.ready().then(() => {
     statusBar.styleDefault();
     splashScreen.hide();
-
-    // OneSignal Code start:
-    // Enable to debug issues:
-    // window["plugins"].OneSignal.setLogLevel({logLevel: 4, visualLevel: 4});
-
-    var notificationOpenedCallback = function(jsonData) {
-      console.log('notificationOpenedCallback: ' + JSON.stringify(jsonData));
-    };
-
-    this.oneSignal.startInit('de69c52c-08b7-4984-8ecf-9f3eec316948', '703322744261');
-
-    this.oneSignal.inFocusDisplaying(this.oneSignal.OSInFocusDisplayOption.InAppAlert);
-    
-    this.oneSignal.handleNotificationReceived().subscribe(data => {
-     // do something when notification is received
-     alert(data.payload.title);
-    });
-    
-    this.oneSignal.handleNotificationOpened().subscribe(data => {
-      // do something when a notification is opened
-      alert(data.notification.payload.body);
-    });
-    
-    this.oneSignal.endInit();
-
-    // window["plugins"].OneSignal
-    //   .startInit("de69c52c-08b7-4984-8ecf-9f3eec316948", "YOUR_GOOGLE_PROJECT_NUMBER_IF_ANDROID")
-    //   .handleNotificationOpened(notificationOpenedCallback)
-    //   .endInit();
   });
 }
